Fix concatenate with a nil value when give a warn
Luarocks tries to detect a Lua interpreter version installed in a
system. When this attempt fails it gives a warning that built or
installed modules may be configured improperly (say, configured for 5.1,
but intended to be run with 5.3).

If there are no --lua-ver[sion] and --lua-dir command-line options or
corresponding configuration values luarocks still tries to find a Lua
interpreter, but fails to give a warning due to concatenation a string
with `nil`. The error looks so:

> /usr/share/tarantool/luarocks/cmd.lua:392: attempt to concatenate
> local 'lua_version' (a nil value)'

The patch changes a warning to report a Lua version from a configuration
when the command line options are not passed.

See https://github.com/luarocks/luarocks/issues/961